@@ -11,25 +14,32 @@ android {
     signingConfigs {
         release
     }
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
     testOptions {
-        unitTests.returnDefaultValues = true
+        unitTests {
+            returnDefaultValues = true
+            all {
+                useJUnit {
+                    excludeCategories 'de.westnordost.streetcomplete.IntegrationTests'
+                }
+            }
+        }
     }
 
     defaultConfig {
         applicationId "de.westnordost.streetcomplete"
         minSdkVersion 17
-        targetSdkVersion 27
-        versionCode 901
-        versionName "9.0"
-        testInstrumentationRunner "android.test.InstrumentationTestRunner"
+        targetSdkVersion 28
+        versionCode 1003
+        versionName "10.0"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         buildConfigField "String", "MAPZEN_API_KEY", MAPZEN_API_KEY
     }
     buildTypes {
 
         release {
             minifyEnabled true
+            useProguard false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
             signingConfig signingConfigs.release
         }
