@@ -1,6 +1,6 @@
 package de.westnordost.streetcomplete.data.osm.changes;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
 import java.util.HashMap;
