@@ -1,7 +1,7 @@
 package de.westnordost.streetcomplete.data.osm.download;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import de.westnordost.streetcomplete.data.osm.ElementGeometry;
 import de.westnordost.osmapi.map.data.Element;
