@@ -8,7 +8,7 @@ import android.hardware.SensorManager;
 import android.location.Location;
 import android.os.Handler;
 import android.os.HandlerThread;
-import android.support.annotation.AnyThread;
+import androidx.annotation.AnyThread;
 import android.view.Display;
 import android.view.Surface;
 
