@@ -1,6 +1,5 @@
 package de.westnordost.streetcomplete.tangram;
 
-import android.animation.ObjectAnimator;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
@@ -9,11 +8,11 @@ import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.AnyThread;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.annotation.AnyThread;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -88,7 +87,7 @@ public class MapControlsFragment extends Fragment
 		View view = inflater.inflate(R.layout.fragment_map_controls, container, false);
 		compassNeedle = view.findViewById(R.id.compassNeedle);
 
-		view.findViewById(R.id.compass).setOnClickListener(v ->
+		view.findViewById(R.id.compassView).setOnClickListener(v ->
 		{
 			boolean isFollowing = mapFragment.isFollowingPosition();
 			boolean isCompassMode = mapFragment.isCompassMode();
@@ -158,7 +157,7 @@ public class MapControlsFragment extends Fragment
 		return view;
 	}
 
-	@Override public void onViewCreated(View view, @Nullable Bundle savedInstanceState)
+	@Override public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState)
 	{
 		super.onViewCreated(view, savedInstanceState);
 		ViewUtils.postOnLayout(view, () ->
