@@ -9,19 +9,18 @@ import android.graphics.drawable.Drawable;
 import android.hardware.SensorManager;
 import android.location.Location;
 import android.os.Bundle;
-import android.support.annotation.AnyThread;
-import android.support.annotation.CallSuper;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.preference.PreferenceManager;
+import androidx.annotation.AnyThread;
+import androidx.annotation.CallSuper;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.preference.PreferenceManager;
 import android.text.Html;
 import android.text.TextUtils;
 import android.text.method.LinkMovementMethod;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.ViewTreeObserver;
 import android.view.WindowManager;
 import android.widget.TextView;
 
@@ -178,7 +177,8 @@ public class MapFragment extends Fragment implements
 			initMarkers();
 			followPosition();
 			showLocation();
-			ViewUtils.postOnLayout(getView(), this::updateView);
+			View v = getView();
+			if(v != null) ViewUtils.postOnLayout(v, this::updateView);
 		}
 	}
 
@@ -355,6 +355,7 @@ public class MapFragment extends Fragment implements
 
 	protected void updateView()
 	{
+		if(controller == null) return;
 		updateAccuracy();
 		if(shouldCenterCurrentPosition())
 		{
