@@ -1,6 +1,10 @@
 package de.westnordost.streetcomplete.data.osm.persist;
 
+import org.junit.Before;
+import org.junit.Test;
+
 import static org.assertj.core.api.Assertions.*;
+import static org.junit.Assert.*;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -34,21 +38,21 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 	private ElementGeometryDao geometryDao;
 	private OsmQuestDao dao;
 
-	@Override public void setUp() throws Exception
+	@Before public void createDaos()
 	{
-		super.setUp();
 		geometryDao = new ElementGeometryDao(dbHelper, serializer);
-		List<QuestType> list = new ArrayList<>();
-		list.add(new TestQuestType());
-		list.add(new TestQuestType2());
-		list.add(new TestQuestType3());
-		list.add(new TestQuestType4());
-		list.add(new TestQuestType5());
+		List<QuestType> list = Arrays.asList(
+			new TestQuestType(),
+			new TestQuestType2(),
+			new TestQuestType3(),
+			new TestQuestType4(),
+			new TestQuestType5()
+		);
 
 		dao = new OsmQuestDao(dbHelper, serializer, new QuestTypeRegistry(list));
 	}
 
-	public void testAddGetNoChanges()
+	@Test public void addGetNoChanges()
 	{
 		OsmQuest quest = createNewQuest(11, Element.Type.NODE);
 
@@ -60,7 +64,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		checkEqual(quest, dbQuest);
 	}
 
-	public void testAddGetWithChanges()
+	@Test public void addGetWithChanges()
 	{
 		List<StringMapEntryChange> changes = new ArrayList<>();
 		changes.add(new StringMapEntryAdd("a key", "a value"));
@@ -77,7 +81,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		checkEqual(quest, dbQuest);
 	}
 
-	public void testGetAllByBBox()
+	@Test public void getAllByBBox()
 	{
 		OsmQuest quest1 = createNewQuest(11, Element.Type.NODE);
 		OsmQuest quest2 = createNewQuest(12, Element.Type.NODE, new ElementGeometry(new OsmLatLon(11,11)));
@@ -88,7 +92,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		assertEquals(2,dao.getAll(null, null, null, null, null).size());
 	}
 
-	public void testGetAllByElementTypeAndId()
+	@Test public void getAllByElementTypeAndId()
 	{
 		OsmQuest quest1 = createNewQuest(11, Element.Type.NODE);
 		OsmQuest quest2 = createNewQuest(12, Element.Type.WAY);
@@ -99,7 +103,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		assertEquals(1,dao.getAll(null, null, null, Element.Type.WAY, 12L).size());
 	}
 
-	public void testGetAllByMultipleQuestTypes()
+	@Test public void getAllByMultipleQuestTypes()
 	{
 		ElementGeometry geom = new ElementGeometry(new OsmLatLon(5,5));
 
@@ -144,7 +148,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		}
 	}
 
-	public void testDeleteAllClosed()
+	@Test public void deleteAllClosed()
 	{
 		OsmQuest quest1 = createNewQuest(1, Element.Type.NODE);
 		quest1.setStatus(QuestStatus.CLOSED);
@@ -157,7 +161,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		assertEquals(2,dao.deleteAllClosed(System.currentTimeMillis() + 10000L));
 	}
 
-	public void testDeleteAllUnsolved()
+	@Test public void deleteAllUnsolved()
 	{
 		OsmQuest quest1 = createNewQuest(1, Element.Type.NODE);
 		quest1.setStatus(QuestStatus.NEW);
@@ -171,7 +175,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		assertEquals(2,dao.deleteAllUnsolved(System.currentTimeMillis() + 10000L));
 	}
 
-	public void testDeleteReverted()
+	@Test public void dDeleteReverted()
 	{
 		ElementGeometry geom = new ElementGeometry(new OsmLatLon(5,5));
 
@@ -189,7 +193,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		assertEquals(1,dao.deleteAllReverted(Element.Type.NODE,1));
 	}
 
-	public void testGetNextNewAt()
+	@Test public void getNextNewAt()
 	{
 		ElementGeometry geom = new ElementGeometry(new OsmLatLon(5,5));
 
@@ -208,7 +212,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		assertEquals(nextQuest.getType().getClass().getSimpleName(),"TestQuestType4");
 	}
 
-	public void testGetNoFittingNextNewAt()
+	@Test public void getNoFittingNextNewAt()
 	{
 		ElementGeometry geom = new ElementGeometry(new OsmLatLon(5,5));
 
@@ -220,7 +224,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		assertNull(dao.getNextNewAt(1,Arrays.asList("TestQuestType")));
 	}
 
-	public void testGetNoNextNewAt()
+	@Test public void getNoNextNewAt()
 	{
 		ElementGeometry geom = new ElementGeometry(new OsmLatLon(5,5));
 
@@ -232,7 +236,7 @@ public class OsmQuestDaoTest extends ApplicationDbTestCase
 		assertNull(dao.getNextNewAt(1,null));
 	}
 
-	public void testGetAllIds()
+	@Test public void getAllIds()
 	{
 		ElementGeometry geom = new ElementGeometry(new OsmLatLon(5,5));
 
