@@ -1,5 +1,8 @@
 package de.westnordost.streetcomplete.data.osm.persist;
 
+import org.junit.Before;
+import org.junit.Test;
+
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
@@ -8,18 +11,19 @@ import de.westnordost.streetcomplete.data.ApplicationDbTestCase;
 import de.westnordost.osmapi.map.data.OsmWay;
 import de.westnordost.osmapi.map.data.Way;
 
+import static org.junit.Assert.assertEquals;
+
 
 public class WayDaoTest extends ApplicationDbTestCase
 {
 	private WayDao dao;
 
-	@Override public void setUp() throws Exception
+	@Before public void createDao()
 	{
-		super.setUp();
 		dao = new WayDao(dbHelper, serializer);
 	}
 
-	public void testPutGetNoTags()
+	@Test public void putGetNoTags()
 	{
 		Way way = new OsmWay(5, 1, Arrays.asList(1L,2L,3L,4L), null);
 		dao.put(way);
@@ -28,7 +32,7 @@ public class WayDaoTest extends ApplicationDbTestCase
 		checkEqual(way, dbWay);
 	}
 
-	public void testPutGetWithTags()
+	@Test public void putGetWithTags()
 	{
 		Map<String,String> tags = new HashMap<>();
 		tags.put("a key", "a value");
