@@ -3,21 +3,25 @@ package de.westnordost.streetcomplete.data.tiles;
 import android.graphics.Point;
 import android.graphics.Rect;
 
+import org.junit.Before;
+import org.junit.Test;
+
 import java.util.List;
 
 import de.westnordost.streetcomplete.data.ApplicationDbTestCase;
+
+import static org.junit.Assert.*;
 
 public class DownloadedTilesDaoTest extends ApplicationDbTestCase
 {
 	private DownloadedTilesDao dao;
 
-	@Override public void setUp() throws Exception
+	@Before public void createDao()
 	{
-		super.setUp();
 		dao = new DownloadedTilesDao(dbHelper);
 	}
 
-	public void testPutGetOne()
+	@Test public void putGetOne()
 	{
 		dao.put(new Rect(5,8,5,8), "Huhu");
 		List<String> huhus = dao.get(new Rect(5,8,5,8),0);
@@ -26,14 +30,14 @@ public class DownloadedTilesDaoTest extends ApplicationDbTestCase
 		assertTrue(huhus.contains("Huhu"));
 	}
 
-	public void testPutGetOld()
+	@Test public void putGetOld()
 	{
 		dao.put(new Rect(5,8,5,8), "Huhu");
 		List<String> huhus = dao.get(new Rect(5,8,5,8),System.currentTimeMillis() + 1000);
 		assertTrue(huhus.isEmpty());
 	}
 
-	public void testPutSomeOld() throws InterruptedException
+	@Test public void putSomeOld() throws InterruptedException
 	{
 		dao.put(new Rect(0,0,1,3), "Huhu");
 		Thread.sleep(2000);
@@ -42,20 +46,20 @@ public class DownloadedTilesDaoTest extends ApplicationDbTestCase
 		assertTrue(huhus.isEmpty());
 	}
 
-	public void testPutMoreGetOne()
+	@Test public void putMoreGetOne()
 	{
 		dao.put(new Rect(5,8,6,10), "Huhu");
 		assertFalse(dao.get(new Rect(5,8,5,8),0).isEmpty());
 		assertFalse(dao.get(new Rect(6,10,6,10),0).isEmpty());
 	}
 
-	public void testPutOneGetMore()
+	@Test public void putOneGetMore()
 	{
 		dao.put(new Rect(5,8,5,8), "Huhu");
 		assertTrue(dao.get(new Rect(5,8,5,9),0).isEmpty());
 	}
 
-	public void testRemove()
+	@Test public void remove()
 	{
 		dao.put(new Rect(0,0,3,3), "Huhu");
 		dao.put(new Rect(0,0,0,0), "Haha");
@@ -63,7 +67,7 @@ public class DownloadedTilesDaoTest extends ApplicationDbTestCase
 		assertEquals(2, dao.remove(new Point(0,0))); // removes huhu, haha at 0,0
 	}
 
-	public void testPutSeveralQuestTypes()
+	@Test public void putSeveralQuestTypes()
 	{
 		dao.put(new Rect(0,0,5,5), "Huhu");
 		dao.put(new Rect(4,4,6,6), "hoho");
