@@ -1,44 +1,49 @@
 package de.westnordost.streetcomplete.data.visiblequests;
 
+import org.junit.Before;
+import org.junit.Test;
+
 import de.westnordost.streetcomplete.data.ApplicationDbTestCase;
 import de.westnordost.streetcomplete.data.osm.persist.test.DisabledTestQuestType;
 import de.westnordost.streetcomplete.data.osm.persist.test.TestQuestType;
 
+import static org.junit.Assert.*;
+
 public class VisibleQuestTypeDaoTest extends ApplicationDbTestCase
 {
 	private VisibleQuestTypeDao dao;
+
 	private TestQuestType testQuestType = new TestQuestType();
 	private DisabledTestQuestType disabledTestQuestType = new DisabledTestQuestType();
 
-	@Override public void setUp() throws Exception
+	@Before public void createDao()
 	{
-		super.setUp();
 		dao = new VisibleQuestTypeDao(dbHelper);
 	}
 
-	public void testDefaultEnabledQuest()
+	@Test public void defaultEnabledQuest()
 	{
 		assertTrue(dao.isVisible(testQuestType));
 	}
 
-	public void testDefaultDisabledQuests()
+	@Test public void defaultDisabledQuests()
 	{
 		assertFalse(dao.isVisible(disabledTestQuestType));
 	}
 
-	public void testDisableQuest()
+	@Test public void disableQuest()
 	{
 		dao.setVisible(testQuestType, false);
 		assertFalse(dao.isVisible(testQuestType));
 	}
 
-	public void testEnableQuest()
+	@Test public void enableQuest()
 	{
 		dao.setVisible(disabledTestQuestType, true);
 		assertTrue(dao.isVisible(disabledTestQuestType));
 	}
 
-	public void testReset()
+	@Test public void reset()
 	{
 		dao.setVisible(testQuestType, false);
 		assertFalse(dao.isVisible(testQuestType));
