@@ -1,6 +1,6 @@
 package de.westnordost.streetcomplete.data.osmnotes;
 
-import java.util.ArrayList;
+import java.util.List;
 
 import de.westnordost.osmapi.map.data.Element;
 import de.westnordost.osmapi.map.data.LatLon;
@@ -13,7 +13,7 @@ public class CreateNote
 	public LatLon position;
 	public Element.Type elementType;
 	public Long elementId;
-	public ArrayList<String> imagePaths;
+	public List<String> imagePaths;
 
 	public boolean hasAssociatedElement()
 	{
