@@ -1,11 +1,12 @@
 package de.westnordost.streetcomplete.data.osmnotes;
 
 import android.graphics.Point;
+import androidx.annotation.Nullable;
 
-import java.util.ArrayList;
+import java.util.List;
 
 public interface CreateNoteListener
 {
 	/** Called when the user wants to leave a note which is not related to a quest */
-	void onLeaveNote(String note, ArrayList<String> imagePaths, Point screenPosition);
+	void onLeaveNote(String note, @Nullable List<String> imagePaths, Point screenPosition);
 }
