@@ -28,7 +28,7 @@ import de.westnordost.osmapi.user.User;
 import de.westnordost.streetcomplete.quests.opening_hours.model.TimeRange;
 import de.westnordost.streetcomplete.quests.opening_hours.model.Weekdays;
 import de.westnordost.streetcomplete.quests.localized_name.LocalizedName;
-import de.westnordost.streetcomplete.quests.postbox_collection_times.WeekdaysTimes;
+import de.westnordost.streetcomplete.quests.postbox_collection_times.WeekdaysTimesRow;
 
 @Singleton
 public class KryoSerializer implements Serializer
@@ -57,7 +57,7 @@ public class KryoSerializer implements Serializer
 			OpeningMonthsRow.class,
 			OpeningWeekdaysRow.class,
 			LocalizedName.class,
-			WeekdaysTimes.class,
+			WeekdaysTimesRow.class,
 			OsmLatLon.class,
 	};
 
